# docs/openapi/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "不動産投資意思決定支援 API",
        "title": "Yield-Guard API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/land-appraisals": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "land-prices"
                ],
                "summary": "地価公示統計",
                "parameters": [
                    {
                        "type": "string",
                        "description": "都道府県コード (例: 13)",
                        "name": "area",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "公示年 (2022-2030)",
                        "name": "year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "市区町村コード",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用途区分 (00=住宅地, 05=商業地, 07=準工業地, 09=工業地)",
                        "name": "division",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.AppraisalComparisonResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/land-prices/compare": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "land-prices"
                ],
                "summary": "土地価格比較",
                "parameters": [
                    {
                        "type": "string",
                        "description": "都道府県コード (例: 13)",
                        "name": "area",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "開始年 (2005以降)",
                        "name": "year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "開始四半期 (1-4)",
                        "name": "quarter",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "終了年",
                        "name": "to_year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "終了四半期 (1-4)",
                        "name": "to_quarter",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "検討物件価格 (円)",
                        "name": "price",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "市区町村コード",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "土地面積 (m²)",
                        "name": "area_sqm",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.LandPriceComparison"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/land-prices/estimate": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "land-prices"
                ],
                "summary": "理論価格推定",
                "parameters": [
                    {
                        "type": "string",
                        "description": "都道府県コード (例: 13)",
                        "name": "area",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "開始年 (2005以降)",
                        "name": "year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "開始四半期 (1-4)",
                        "name": "quarter",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "終了年",
                        "name": "to_year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "終了四半期 (1-4)",
                        "name": "to_quarter",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "物件価格 (円)",
                        "name": "price",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "土地面積 (m²)",
                        "name": "area_sqm",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "市区町村コード",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "築年数",
                        "name": "building_age",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "最寄り駅徒歩分",
                        "name": "station_minutes",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "需要スコア (A-E)",
                        "name": "ridership_score",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.TheoreticalPriceResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/land-prices/stats": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "land-prices"
                ],
                "summary": "土地取引価格統計",
                "parameters": [
                    {
                        "type": "string",
                        "description": "都道府県コード (例: 13)",
                        "name": "area",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "開始年 (2005以降)",
                        "name": "year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "開始四半期 (1-4)",
                        "name": "quarter",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "終了年",
                        "name": "to_year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "終了四半期 (1-4)",
                        "name": "to_quarter",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "市区町村コード",
                        "name": "city",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.LandPriceStats"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/municipalities": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "land-prices"
                ],
                "summary": "市区町村一覧",
                "parameters": [
                    {
                        "type": "string",
                        "description": "都道府県コード (例: 13)",
                        "name": "area",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/mlit.Municipality"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.AppraisalComparisonResult": {
            "type": "object",
            "properties": {
                "appraisalCount": {
                    "description": "標準地点数",
                    "type": "integer"
                },
                "appraisalMedianPerSqm": {
                    "description": "公示価格中央値（円/m²）",
                    "type": "number"
                },
                "appraisalTrend": {
                    "description": "平均変動率（小数）",
                    "type": "number"
                },
                "trendLabel": {
                    "description": "\"上昇\" / \"安定\" / \"下落\"",
                    "type": "string"
                }
            }
        },
...
        "domain.LandPriceComparison": {
            "type": "object",
            "properties": {
                "assessment": {
                    "description": "\"割安\" / \"相場\" / \"割高\"",
                    "type": "string"
                },
                "diffFromAverage": {
                    "type": "number"
                },
                "diffFromMedian": {
                    "type": "number"
                },
                "inputArea": {
                    "type": "number"
                },
                "inputLandPrice": {
                    "type": "number"
                },
                "inputPricePerTsubo": {
                    "type": "number"
                },
                "stats": {
                    "$ref": "#/definitions/domain.LandPriceStats"
                }
            }
        },
        "domain.LandPriceStats": {
            "type": "object",
            "properties": {
                "averageTsubo": {
                    "type": "number"
                },
                "count": {
                    "type": "integer"
                },
                "lowDataWarning": {
                    "description": "件数 \u003c 10 件時 true",
                    "type": "boolean"
                },
                "maxTsubo": {
                    "type": "number"
                },
                "medianTsubo": {
                    "type": "number"
                },
                "minTsubo": {
                    "type": "number"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.LandTransaction"
                    }
                },
                "urbanRisks": {
                    "description": "都市計画リスク一覧",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.UrbanRisk"
                    }
                },
                "warningMessage": {
                    "type": "string"
                },
                "zoning": {
                    "description": "取引データから抽出した用途地域情報",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.ZoningSummary"
                        }
                    ]
                }
            }
        },
        "domain.LandTransaction": {
            "type": "object",
            "properties": {
                "area": {
                    "type": "number"
                },
                "buildingCoverage": {
                    "type": "string"
                },
                "buildingYear": {
                    "description": "建築年（西暦）",
                    "type": "integer"
                },
                "cityPlanning": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "floorAreaRatio": {
                    "type": "string"
                },
                "period": {
                    "type": "string"
                },
                "pricePerSqm": {
                    "type": "number"
                },
                "pricePerTsubo": {
                    "type": "number"
                },
                "stationMinutes": {
                    "description": "最寄り駅徒歩分",
                    "type": "integer"
                },
                "tradePrice": {
                    "type": "number"
                }
            }
        },
...
        "domain.RidershipDemandScore": {
            "type": "string",
            "enum": [
                "A",
                "B",
                "C",
                "D",
                "E"
            ],
            "x-enum-comments": {
                "RidershipScoreA": "\u003e=100,000人/日",
                "RidershipScoreB": "\u003e=30,000人/日",
                "RidershipScoreC": "\u003e=10,000人/日",
                "RidershipScoreD": "\u003e=2,000人/日",
                "RidershipScoreE": "\u003c2,000人/日"
            },
            "x-enum-descriptions": [
                "\u003e=100,000人/日",
                "\u003e=30,000人/日",
                "\u003e=10,000人/日",
                "\u003e=2,000人/日",
                "\u003c2,000人/日"
            ],
            "x-enum-varnames": [
                "RidershipScoreA",
                "RidershipScoreB",
                "RidershipScoreC",
                "RidershipScoreD",
                "RidershipScoreE"
            ]
        },
...
        "domain.TheoreticalPriceResult": {
            "type": "object",
            "properties": {
                "ageCorrection": {
                    "description": "築年数補正 (-0.3〜0.3)",
                    "type": "number"
                },
                "deviationPct": {
                    "description": "乖離率 % (正=割高, 負=割安)",
                    "type": "number"
                },
                "hasRidershipData": {
                    "description": "乗降客数補正が有効か",
                    "type": "boolean"
                },
                "hasStationData": {
                    "description": "駅距離補正が有効か",
                    "type": "boolean"
                },
                "isLowDataWarning": {
                    "description": "築年数サンプル不足",
                    "type": "boolean"
                },
                "medianBuildingAge": {
                    "description": "取引事例中央値築年数",
                    "type": "integer"
                },
                "medianStationMinutes": {
                    "description": "取引事例中央値駅距離(分)",
                    "type": "integer"
                },
                "ridershipCorrection": {
                    "description": "乗降客数補正 (-0.15〜0.15)",
                    "type": "number"
                },
                "ridershipScore": {
                    "description": "需要スコア (有効時のみ)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.RidershipDemandScore"
                        }
                    ]
                },
                "stationCorrection": {
                    "description": "駅距離補正 (-0.2〜0.2)",
                    "type": "number"
                },
                "theoreticalPriceJPY": {
                    "description": "理論価格 (円)",
                    "type": "number"
                }
            }
        },
        "domain.UrbanRisk": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "リスクコード（例: URBANIZATION_CONTROL）",
                    "type": "string"
                },
                "description": {
                    "description": "詳細説明",
                    "type": "string"
                },
                "level": {
                    "description": "ERROR / WARNING / INFO",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.UrbanRiskLevel"
                        }
                    ]
                },
                "title": {
                    "description": "短いタイトル",
                    "type": "string"
                }
            }
        },
        "domain.UrbanRiskLevel": {
            "type": "string",
            "enum": [
                "ERROR",
                "WARNING",
                "INFO"
            ],
            "x-enum-varnames": [
                "UrbanRiskLevelError",
                "UrbanRiskLevelWarning",
                "UrbanRiskLevelInfo"
            ]
        },
...
        "domain.ZoningSummary": {
            "type": "object",
            "properties": {
                "buildingCoverage": {
                    "description": "最頻の建ぺい率（例: 60%）",
                    "type": "string"
                },
                "cityPlanning": {
                    "description": "最頻の都市計画区域（例: 第一種住居地域）",
                    "type": "string"
                },
                "floorAreaRatio": {
                    "description": "最頻の容積率（例: 200%）",
                    "type": "string"
                }
            }
        },
        "mlit.Municipality": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "市区町村コード (例: \"13101\")",
                    "type": "string"
                },
                "name": {
                    "description": "市区町村名 (例: \"千代田区\")",
                    "type": "string"
                }
            }
        }
    }
}